個別に記録できるところまで

+++ db/migrate/20240922111850_create_temperatures.rb
@@ -1,7 +1,7 @@
 class CreateTemperatures < ActiveRecord::Migration[7.0]
   def change
     create_table :temperatures do |t|
-      t.integer :netu
+      t.decimal :netu
 
       t.timestamps
     end
